more testing on shawn (bin tree works have to correct the lqi/rssi relation on backoff)

+++ wiselib.testing/algorithms/tracking/PLTT_target.h
@@ -376,7 +376,7 @@ namespace wiselib
 		PrivacyMessage randomize_privacy_message;
 #endif
 #ifdef CONFIG_PLTT_TARGET_H_MINI_RUN
-		uint8_t target_mini_run_times;
+		uint32_t target_mini_run_times;
 #endif
 		uint8_t status;
 	};
